@@ -34,40 +34,41 @@
     "npm": "Please use yarn instead of NPM to install dependencies"
   },
   "dependencies": {
-    "@angular/animations": "~13.3.11",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
-    "@angular/youtube-player": "^13.3.9",
+    "@angular/animations": "^14.2.9",
+    "@angular/cdk": "^14.2.6",
+    "@angular/common": "^14.2.9",
+    "@angular/compiler": "^14.2.9",
+    "@angular/core": "^14.2.9",
+    "@angular/forms": "^14.2.9",
+    "@angular/material": "^14.2.6",
+    "@angular/platform-browser": "^14.2.9",
+    "@angular/platform-browser-dynamic": "^14.2.9",
+    "@angular/router": "^14.2.9",
+    "@angular/youtube-player": "^14.2.6",
     "@mapbox/mapbox-gl-geocoder": "^5.0.0",
     "angular-tag-cloud-module": "^13.0.0",
     "core-js": "3.22.4",
     "lodash-es": "^4.17.20",
     "mapbox-gl": "^2.10.0",
     "ngx-logger": "^5.0.11",
-    "ngx-mapbox-gl": "^7.1.2",
-    "ngx-markdown": "^13",
+    "ngx-mapbox-gl": "^9.1.0",
+    "ngx-markdown": "^14",
     "ngx-web-storage": "^0.2.0",
     "rxjs": "7.5.6",
+    "ts-jest": "^29",
     "tslib": "^2.3.1",
     "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular-eslint/builder": "13.5.0",
-    "@angular-eslint/eslint-plugin": "13.5.0",
-    "@angular-eslint/eslint-plugin-template": "13.5.0",
-    "@angular-eslint/schematics": "13.5.0",
-    "@angular-eslint/template-parser": "13.5.0",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.8",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.8",
+    "@angular/compiler-cli": "^14.2.9",
+    "@angular/language-service": "^14.2.9",
     "@types/date-fns": "^2.6.0",
     "@types/jest": "^29.0.0",
     "@types/mapbox-gl": "^2.7.5",
@@ -75,9 +76,8 @@
     "@typescript-eslint/eslint-plugin": "^5.36.2",
     "@typescript-eslint/parser": "^5.36.2",
     "eslint": "^8.23.0",
-    "jest": "^27",
-    "jest-preset-angular": "11",
-    "ts-jest": "^27",
-    "typescript": "4.6.2"
+    "jest": "^28",
+    "jest-preset-angular": "12",
+    "typescript": "4.7.2"
   }
 }
