@@ -4,7 +4,9 @@ import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
 import {ActivatedRoute} from '@angular/router';
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {EnvironmentService} from '@shared/services/environment.service';
-import {Feature, GeoJSON, Point} from 'geojson';
+// Since update to angular 14:./src/app/map/map.component.ts:4:0-41 - Error: Module not found: Error: Can't resolve 'geojson' in '/Users/tillkuhn/git/angkor/angular/src/app/map'
+// new from location is geojson -> @mapbox/geojson-types
+import {Feature, GeoJSON, Point} from '@mapbox/geojson-types';
 import {MapComponent as MapboxGLMapComponent} from 'ngx-mapbox-gl';
 import {MapboxGeoJSONFeature, MapLayerMouseEvent, Marker} from 'mapbox-gl';
 import {MasterDataService} from '@shared/services/master-data.service';
